+++ app/src/main/java/org/toop/app/canvas/TicTacToeBitCanvas.java
@@ -3,10 +3,6 @@ package org.toop.app.canvas;
 import javafx.scene.paint.Color;
 import org.toop.app.App;
 import org.toop.framework.gameFramework.model.game.TurnBasedGame;
-import org.toop.game.games.tictactoe.BitboardTicTacToe;
-
-import java.util.Arrays;
-import java.util.function.Consumer;
 
 public class TicTacToeBitCanvas extends BitGameCanvas{
     public TicTacToeBitCanvas() {
@@ -29,8 +25,8 @@ public class TicTacToeBitCanvas extends BitGameCanvas{
     }
 
     private void drawMoves(long[] gameBoard){
-        loopOverBoard(gameBoard[0], (i) -> drawX(Color.RED, (int)i));
-        loopOverBoard(gameBoard[1], (i) -> drawO(Color.BLUE, (Integer) i));
+        loopOverBoard(gameBoard[0], (i) -> drawX(Color.RED, i));
+        loopOverBoard(gameBoard[1], (i) -> drawO(Color.BLUE, i));
 
     }
